relatorio_clientes_cashu/requirements.txt
streamlit
snowflake-connector-python[pandas]
snowflake-snowpark-python
pandas
plotly
itables
python-dotenv
